+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+ - People were confused by our link sharing, we've finally improved the way they look when shared.
+
 ## [1.10.2] - 2021-01-06
 
  - Minor fixes.